# ENH compute histograms only for allowed features in HGBT