#### Reference Issues/PRs
Follow-up of #21020.

#### What does this implement/fix? Explain your changes.
This PR restricts the computation of histograms in `HistGradientBoostingRegressor` and `HistGradientBoostingClassifier` to features that are allowed to be split on. This gives a small boost in performance (fit time).

#### Any other comments?